move build section to its file
